+++ RELEASE-NOTES.md
@@ -9,6 +9,7 @@
 #### Additions
 
 * Created `StatementListProvider` interface
+* Created `ConstraintFactory` class
 
 ## Version 0.1 (2014-09-21)